+++ src/core/unitTest/index.ts
@@ -10,6 +10,7 @@ import type { IUnitTestConfig, IUnitTestState } from './types';
 export type { ICdsUnitTestConfig, ICdsUnitTestState } from './AdtCdsUnitTest';
 export { AdtCdsUnitTest } from './AdtCdsUnitTest';
 export { AdtUnitTest } from './AdtUnitTest';
+export { AdtUnitTestLegacy } from './AdtUnitTestLegacy';
 export type { IUnitTestConfig, IUnitTestState } from './types';
 
 // Type alias for AdtUnitTest